requirements.txt
streamlit
pandas
requests
plotly
python-dateutil
openpyxl
python-dotenv
matplotlib
